add java judge worker, need testing with test cases.

+++ src/main/java/plus/justice/models/database/Submission.java
@@ -25,10 +25,10 @@ public class Submission {
     private Long id;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "problem_id")
-    private Integer problemId;
+    private Long problemId;
 
     @Column(name = "language")
     private Integer language;
@@ -64,19 +64,19 @@ public class Submission {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getProblemId() {
+    public Long getProblemId() {
         return problemId;
     }
 
-    public void setProblemId(Integer problemId) {
+    public void setProblemId(Long problemId) {
         this.problemId = problemId;
     }
